+++ repo_update/exceptions.py
@@ -1,0 +1,6 @@
+class RepoUpdateException(Exception):
+    pass
+
+
+class KPIReadException(RepoUpdateException):
+    pass
