+++ requirements.txt
@@ -1,0 +1,1 @@
+PyYAML==3.11
